defining helpers
