+++ packages/ts-components/src/components/social-embed/SocialMediaEmbed.tsx
@@ -9,10 +9,13 @@ import {
   CustomIconContainer,
   Header
 } from './styles';
-import get from 'lodash.get';
 import { InfoIcon } from '../inline-message/InfoIcon';
+import { openPrivacyModal, ModalType } from './helpers/privacyModal';
 // @ts-ignore
 import InteractiveWrapper from '@times-components/interactive-wrapper';
+import { checkVendorConsent } from './helpers/vendorConsent';
+// import { getVendorTitle } from './helpers/getVendorTitle';
+import { enableCookies } from './helpers/enableCookies';
 
 export declare type SocialEmbedProps = {
   element: any;
@@ -38,35 +41,11 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
   const [allowedOnce, setAllowedOnce] = useState(false);
   const [isSocialAllowed, setIsSocialAllowed] = useState(false);
 
-  const checkVendorConsent = () => {
-    if (window.__tcfapi) {
-      window.__tcfapi('getCustomVendorConsents', 2, (data, success) => {
-        if (success && data && data.consentedVendors) {
-          const isSocialVendorAllowed = data.consentedVendors.some(
-            (vendor: { name: string }) =>
-              vendor.name.toLowerCase() === vendorName.toLowerCase()
-          );
-          setIsSocialAllowed(isSocialVendorAllowed);
-          // tslint:disable-next-line:no-console
-          console.log(
-            `Consent check for ${vendorName}:`,
-            isSocialVendorAllowed
-          );
-        } else {
-          // tslint:disable-next-line:no-console
-          console.log(
-            `Error fetching consent data or ${vendorName} embed not allowed`
-          );
-        }
-      });
-    }
-  };
-
   useEffect(
     () => {
       // tslint:disable-next-line:no-console
       console.log('useEffect enterred');
-      checkVendorConsent();
+      checkVendorConsent(vendorName, setIsSocialAllowed);
     },
     [vendorName, allowedOnce, isSocialAllowed]
   );
@@ -84,22 +63,6 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
     [isSocialAllowed, allowedOnce]
   );
 
-  enum ModalType {
-    GDPR = 'gdpr',
-    CCPA = 'ccpa'
-  }
-
-  const openPrivacyModal = (type: ModalType, messageId: string) => {
-    const loadModal = get(window, `_sp_.${type}.loadPrivacyManagerModal`);
-
-    if (loadModal) {
-      loadModal(messageId);
-    } else {
-      // tslint:disable-next-line:no-console
-      console.warn('Sourcepoint LoadPrivacyManagerModal is not available');
-    }
-  };
-
   const allowCookiesOnce = () => {
     setAllowedOnce(true);
     setIsSocialAllowed(true);
@@ -115,59 +78,12 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
     );
   };
 
-  const socialMediaVendors: {
-    [key: string]: { id: string; status: string };
-  } = {
-    twitter: { id: '5fab0c31a22863611c5f8764', status: 'pending' },
-    youtube: { id: '5e7ac3fae30e7d1bc1ebf5e8', status: 'pending' }
-  };
-
-  const getVendorTitle = (title: string): string => {
-    if (title === 'twitter') {
-      return 'X (Twitter)';
-    }
-    return socialMediaVendors[title].id || title;
-  };
-
-  const enableCookies = (providerName: string) => {
-    const onCustomConsent = (_: any, success: boolean) => {
-      if (success) {
-        setIsSocialAllowed(true);
-        return {
-          ...socialMediaVendors.vendorName,
-          [vendorName]: {
-            ...socialMediaVendors[providerName],
-            status: 'accepted'
-          }
-        };
-      }
-      return null;
-    };
-
-    const vendorId = socialMediaVendors[providerName].id;
-
-    if (window.__tcfapi && vendorId) {
-      window.__tcfapi(
-        'getCustomVendorConsents',
-        2,
-        (data: any, successful: boolean) => {
-          if (successful && data && data.grants[vendorId]) {
-            (window.__tcfapi as any)(
-              'postCustomConsent',
-              2,
-              onCustomConsent,
-              [vendorId],
-              Object.keys(data.grants[vendorId].purposeGrants),
-              []
-            );
-          } else {
-            // tslint:disable-next-line:no-console
-            console.error(`${vendorName} vendor consent not available:`, data);
-          }
-        }
-      );
-    }
-  };
+  // const socialMediaVendors: {
+  //   [key: string]: { id: string; status: string };
+  // } = {
+  //   twitter: { id: '5fab0c31a22863611c5f8764', status: 'pending' },
+  //   youtube: { id: '5e7ac3fae30e7d1bc1ebf5e8', status: 'pending' }
+  // };
 
   // tslint:disable-next-line:no-console
   console.log('allowedOnce', allowedOnce);
@@ -176,7 +92,7 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
   // tslint:disable-next-line:no-console
   console.log('allowedOnce && isSocialAllowed', allowedOnce && isSocialAllowed);
 
-  return isSocialAllowed || allowedOnce ? (
+  return true ? (
     <InteractiveWrapper
       attributes={element.attributes}
       element={element.value}
@@ -189,7 +105,9 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
         <CustomIconContainer>
           <InfoIcon />
         </CustomIconContainer>
-        <Title>{getVendorTitle(vendorName)} content blocked</Title>
+        <Title>
+          Lili content blocked
+        </Title>
       </Header>
       <Paragraph>
         Please enable cookies and other technologies to view this content. You
@@ -198,7 +116,11 @@ export const SocialMediaEmbed: React.FC<SocialEmbedProps> = ({
           privacy manager.
         </LinkPrivacyManager>
       </Paragraph>
-      <EnableButton onClick={() => enableCookies(vendorName)}>
+      <EnableButton
+        onClick={() =>
+          enableCookies(vendorName, setIsSocialAllowed)
+        }
+      >
         Enable cookies
       </EnableButton>
       <AllowButton onClick={allowCookiesOnce}>Allow cookies once</AllowButton>
