+++ packages/ts-components/src/components/social-embed/helpers/privacyModal.ts
@@ -1,0 +1,17 @@
+import get from 'lodash.get';
+
+export enum ModalType {
+  GDPR = 'gdpr',
+  CCPA = 'ccpa'
+}
+
+export const openPrivacyModal = (type: ModalType, messageId: string) => {
+  const loadModal = get(window, `_sp_.${type}.loadPrivacyManagerModal`);
+
+  if (loadModal) {
+    loadModal(messageId);
+  } else {
+    // tslint:disable-next-line:no-console
+    console.warn('Sourcepoint LoadPrivacyManagerModal is not available');
+  }
+};
